# Conflicts:
#	testar/src/org/fruit/monkey/ConfigTags.java
#	testar/src/org/fruit/monkey/Main.java

+++ testar/src/nl/ou/testar/StateModel/Analysis/AnalysisManager.java
@@ -6,19 +6,21 @@ import com.orientechnologies.orient.core.db.ODatabaseSession;
 import com.orientechnologies.orient.core.db.OrientDB;
 import com.orientechnologies.orient.core.db.OrientDBConfig;
 import com.orientechnologies.orient.core.metadata.schema.OType;
+import com.orientechnologies.orient.core.record.ODirection;
 import com.orientechnologies.orient.core.record.OEdge;
 import com.orientechnologies.orient.core.record.OVertex;
 import com.orientechnologies.orient.core.record.impl.ORecordBytes;
 import com.orientechnologies.orient.core.record.impl.OVertexDocument;
 import com.orientechnologies.orient.core.sql.executor.OResult;
 import com.orientechnologies.orient.core.sql.executor.OResultSet;
-import nl.ou.testar.StateModel.Analysis.HttpServer.JettyServer;
 import nl.ou.testar.StateModel.Analysis.Json.Edge;
 import nl.ou.testar.StateModel.Analysis.Json.Element;
 import nl.ou.testar.StateModel.Analysis.Json.Vertex;
 import nl.ou.testar.StateModel.Analysis.Representation.AbstractStateModel;
+import nl.ou.testar.StateModel.Analysis.Representation.ActionViz;
 import nl.ou.testar.StateModel.Analysis.Representation.TestSequence;
 import nl.ou.testar.StateModel.Persistence.OrientDB.Entity.Config;
+import nl.ou.testar.StateModel.Sequence.SequenceVerdict;
 
 import java.io.*;
 import java.text.DateFormat;
@@ -91,6 +93,7 @@ public class AnalysisManager {
                     abstractStateModels.add(abstractStateModel);
                 }
             }
+            resultSet.close();
         }
         return abstractStateModels;
     }
@@ -128,13 +131,120 @@ public class AnalysisManager {
                         nrOfNodes--;
                     }
                 }
+                nodeResultSet.close();
 
                 String sequenceId = (String) getConvertedValue(OType.STRING, sequenceVertex.getProperty("sequenceId"));
                 Date startDateTime = (Date) getConvertedValue(OType.DATETIME, sequenceVertex.getProperty("startDateTime"));
-                sequenceList.add(new TestSequence(sequenceId, DateFormat.getDateTimeInstance().format(startDateTime), String.valueOf(nrOfNodes)));
+
+                // not the best piece of code, but it works for now
+                int verdict;
+                String verdictValue = (String) getConvertedValue(OType.ANY.STRING, sequenceVertex.getProperty("verdict"));
+                if (verdictValue.equals(SequenceVerdict.COMPLETED_SUCCESFULLY.toString())) {
+                    verdict =TestSequence.VERDICT_SUCCESS;
+                }
+                else if (verdictValue.equals(SequenceVerdict.INTERRUPTED_BY_USER.toString())) {
+                    verdict = TestSequence.VERDICT_INTERRUPT_BY_USER;
+                }
+                else if (verdictValue.equals(SequenceVerdict.INTERRUPTED_BY_ERROR.toString())) {
+                    verdict = TestSequence.VERDICT_INTERRUPT_BY_SYSTEM;
+                }
+                else {
+                    verdict = TestSequence.VERDICT_UNKNOWN;
+                }
+
+                // fetch the number of errors that were encountered during the test run
+                String errorStmt = "SELECT COUNT(*) as nr FROM(TRAVERSE out(\"FirstNode\"), out(\"SequenceStep\") FROM ( SELECT FROM TestSequence WHERE sequenceId = :sequenceId)) WHERE @class = \"SequenceNode\" AND containsErrors = true";
+                params = new HashMap<>();
+                params.put("sequenceId", getConvertedValue(OType.STRING, sequenceVertex.getProperty("sequenceId")));
+                OResultSet errorResultSet = db.query(errorStmt, params);
+                int nrOfErrors = 0;
+                if (errorResultSet.hasNext()) {
+                    OResult errorResult = errorResultSet.next();
+                    nrOfErrors = (int)getConvertedValue(OType.INTEGER, errorResult.getProperty("nr"));
+                }
+                errorResultSet.close();
+                TestSequence testSequence = new TestSequence(sequenceId, DateFormat.getDateTimeInstance().format(startDateTime), String.valueOf(nrOfNodes), verdict);
+                testSequence.setNrOfErrors(nrOfErrors);
+
+                sequenceList.add(testSequence);
             }
         }
+        resultSet.close();
         return sequenceList;
+    }
+
+    public List<ActionViz> fetchTestSequence(String sequenceId) {
+        List<ActionViz> visualizations = new ArrayList<>();
+        // fetch the first sequence node belonging to a given test sequence
+        try (ODatabaseSession db = orientDB.open(dbConfig.getDatabase(), dbConfig.getUser(), dbConfig.getPassword())) {
+            String sequenceStmt = "SELECT FROM(TRAVERSE out(\"FirstNode\") FROM ( SELECT FROM TestSequence WHERE sequenceId = :sequenceId)) WHERE @class = \"SequenceNode\"";
+            Map<String, Object> params = new HashMap<>();
+            params.put("sequenceId", sequenceId);
+            OResultSet resultSet = db.query(sequenceStmt, params);
+
+            if (!resultSet.hasNext()) {
+                return visualizations; // no sequence node found
+            }
+
+            OResult nodeResult = resultSet.next();
+            if (!nodeResult.isVertex()) return visualizations;
+            Optional<OVertex> nodeVertexOptional = nodeResult.getVertex();
+            if (!nodeVertexOptional.isPresent()) {
+                return visualizations;
+            }
+
+            // alright, we have a node
+            OVertex nodeVertex = nodeVertexOptional.get();
+            int counterSource = 1;
+            int counterTarget = 2;
+
+            while(true) {
+                // pffff..alright, now that we have our node vertex, we need to get two things:
+                // 1) the concrete state node that is connected to our sequence node, as we need the screenshot
+                // 2) the sequence step going out, for the description
+
+                // first, see if we have another sequence step from this node
+                OEdge sequenceStepEdge = null;
+                for(OEdge edge : nodeVertex.getEdges(ODirection.OUT, "SequenceStep")) {
+                    sequenceStepEdge = edge;
+                    break; // there should at most be one edge
+                }
+
+                if (sequenceStepEdge == null) {
+                    break; // nothing left to do
+                }
+
+                // next, get the vertex that is at the received end of the step edge
+                OVertex targetVertex = sequenceStepEdge.getTo();
+                // now, fetch the concrete states for both the vertices
+                OVertex sourceState = null;
+                OVertex targetState = null;
+                for(OEdge edge : nodeVertex.getEdges(ODirection.OUT, "Accessed")) {
+                    sourceState = edge.getTo();
+                    break; // there should at most be one edge
+                }
+                for(OEdge edge: targetVertex.getEdges(ODirection.OUT, "Accessed")) {
+                    targetState = edge.getTo();
+                    break; // there should at most be one edge
+                }
+                if (sourceState == null || targetState == null) {
+                    return visualizations;
+                }
+
+                String sourceScreenshot = "n" + formatId(sourceState.getIdentity().toString());
+                processScreenShot(sourceState.getProperty("screenshot"), sourceScreenshot, sequenceId);
+                String targetScreenshot = "n" + formatId(targetState.getIdentity().toString());
+                processScreenShot(targetState.getProperty("screenshot"), targetScreenshot, sequenceId);
+                String actionDescription = (String) getConvertedValue(OType.STRING, sequenceStepEdge.getProperty("actionDescription"));
+                ActionViz actionViz = new ActionViz(sourceScreenshot, targetScreenshot, actionDescription, counterSource, counterTarget);
+                visualizations.add(actionViz);
+                nodeVertex = targetVertex;
+                counterSource++;
+                counterTarget++;
+            }
+            resultSet.close();
+            return visualizations;
+        }
     }
 
     /**
@@ -204,22 +314,26 @@ public class AnalysisManager {
         params.put("identifier", modelIdentifier);
         OResultSet resultSet = db.query(stmt, params);
         elements.addAll(fetchNodes(resultSet, "AbstractState", showCompoundGraph ? "AbstractLayer" : null, modelIdentifier));
+        resultSet.close();
 
         // abstract actions
         stmt = "SELECT FROM AbstractAction WHERE modelIdentifier = :identifier";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "AbstractAction"));
+        resultSet.close();
 
         // Black hole class
         stmt = "SELECT FROM (TRAVERSE out() FROM  (SELECT FROM AbstractState WHERE modelIdentifier = :identifier)) WHERE @class = 'BlackHole'";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchNodes(resultSet, "BlackHole", showCompoundGraph ? "AbstractLayer" : null, modelIdentifier));
+        resultSet.close();
 
 
         // unvisited abstract actions
         stmt = "SELECT FROM UnvisitedAbstractAction WHERE modelIdentifier = :identifier";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "UnvisitedAbstractAction"));
+        resultSet.close();
 
         return elements;
     }
@@ -245,11 +359,13 @@ public class AnalysisManager {
         params.put("identifier", modelIdentifier);
         OResultSet resultSet = db.query(stmt, params);
         elements.addAll(fetchNodes(resultSet, "ConcreteState", showCompoundGraph ? "ConcreteLayer" : null, modelIdentifier));
+        resultSet.close();
 
         // concrete actions
         stmt = "SELECT FROM (TRAVERSE in('isAbstractedBy').outE('ConcreteAction') FROM (SELECT FROM AbstractState WHERE modelIdentifier = :identifier)) WHERE @class = 'ConcreteAction'";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "ConcreteAction"));
+        resultSet.close();
 
         return elements;
     }
@@ -275,21 +391,25 @@ public class AnalysisManager {
         params.put("identifier", modelIdentifier);
         OResultSet resultSet = db.query(stmt, params);
         elements.addAll(fetchNodes(resultSet, "TestSequence", showCompoundGraph ? "SequenceLayer" : null, modelIdentifier));
+        resultSet.close();
 
         // sequence nodes
         stmt = "SELECT FROM (TRAVERSE in('isAbstractedBy').in('Accessed') FROM (SELECT FROM AbstractState WHERE modelIdentifier = :identifier)) WHERE @class = 'SequenceNode'";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchNodes(resultSet, "SequenceNode", showCompoundGraph ? "SequenceLayer" : null, modelIdentifier));
+        resultSet.close();
 
         // sequence steps
         stmt = "SELECT FROM (TRAVERSE in('isAbstractedBy').in('Accessed').outE('SequenceStep') FROM (SELECT FROM AbstractState WHERE modelIdentifier = :identifier)) WHERE @class = 'SequenceStep'";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "SequenceStep"));
+        resultSet.close();
 
         // first node
         stmt = "SELECT FROM (TRAVERSE outE('FirstNode') FROM (SELECT FROM TestSequence WHERE modelIdentifier = :identifier)) WHERE @class = 'FirstNode'";
         resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "FirstNode"));
+        resultSet.close();
 
         return elements;
     }
@@ -328,6 +448,7 @@ public class AnalysisManager {
         params.put("identifier", modelIdentifier);
         OResultSet resultSet = db.query(stmt, params);
         elements.addAll(fetchEdges(resultSet, "Accessed"));
+        resultSet.close();
 
         return elements;
     }
@@ -345,11 +466,13 @@ public class AnalysisManager {
             params.put("rid", internalId);
             OResultSet resultSet = db.query(stmt, params);
             elements.addAll(fetchNodes(resultSet, "Widget", null, concreteStateIdentifier));
+            resultSet.close();
 
             // then get the parent/child relationship between the widgets
             stmt = "SELECT FROM isChildOf WHERE in IN(SELECT @RID FROM (TRAVERSE in('isChildOf') FROM (SELECT FROM Widget WHERE @RID = :rid)))";
             resultSet = db.query(stmt, params);
             elements.addAll(fetchEdges(resultSet, "isChildOf"));
+            resultSet.close();
 
             // create a filename
             StringBuilder builder = new StringBuilder(concreteStateIdentifier);
@@ -456,6 +579,9 @@ public class AnalysisManager {
 
         // save the file to disk
         File screenshotFile = new File( screenshotDir, identifier + ".png");
+        if (screenshotFile.exists()) {
+            return;
+        }
         try {
             FileOutputStream outputStream = new FileOutputStream(screenshotFile);
             outputStream.write(recordBytes.toStream());
